types: move DependencyType related types to types workspace
