+++ src/graph/src/graph.ts
@@ -2,10 +2,12 @@ import { getId, joinDepIDTuple, type DepID } from '@vltpkg/dep-id'
 import { error } from '@vltpkg/error-cause'
 import { satisfies } from '@vltpkg/satisfies'
 import { Spec, type SpecOptions } from '@vltpkg/spec'
-import { type Manifest } from '@vltpkg/types'
+import {
+  type Manifest,
+  type DependencyTypeShort,
+} from '@vltpkg/types'
 import { type Monorepo } from '@vltpkg/workspaces'
 import { inspect, type InspectOptions } from 'util'
-import { type DependencyTypeShort } from './dependencies.js'
 import { type Edge } from './edge.js'
 import { lockfileData } from './lockfile/save.js'
 import { Node, type NodeOptions } from './node.js'
