+++ src/graph/src/lockfile/types.ts
@@ -1,7 +1,10 @@
 import { type DepID } from '@vltpkg/dep-id'
 import { type Spec, type SpecOptions } from '@vltpkg/spec'
-import { type Integrity, type Manifest } from '@vltpkg/types'
-import { type DependencyTypeShort } from '../dependencies.js'
+import {
+  type Integrity,
+  type Manifest,
+  type DependencyTypeShort,
+} from '@vltpkg/types'
 import { type Graph } from '../graph.js'
 
 /**
