+++ src/graph/src/node.ts
@@ -8,8 +8,11 @@ import {
 } from '@vltpkg/dep-id'
 import { typeError } from '@vltpkg/error-cause'
 import { type Spec, type SpecOptions } from '@vltpkg/spec'
-import { type Integrity, type Manifest } from '@vltpkg/types'
-import { type DependencyTypeShort } from './dependencies.js'
+import {
+  type Integrity,
+  type Manifest,
+  type DependencyTypeShort,
+} from '@vltpkg/types'
 import { Edge } from './edge.js'
 import { type GraphLike, type NodeLike } from './types.js'
 import { stringifyNode } from './stringify-node.js'
