+++ src/graph/test/ideal/add-nodes.ts
@@ -1,4 +1,5 @@
 import { joinDepIDTuple } from '@vltpkg/dep-id'
+import { type DependencyTypeShort } from '@vltpkg/types'
 import { type PackageInfoClient } from '@vltpkg/package-info'
 import { kCustomInspect, Spec, type SpecOptions } from '@vltpkg/spec'
 import { PathScurry } from 'path-scurry'
@@ -6,7 +7,6 @@ import t from 'tap'
 import {
   type AddImportersDependenciesMap,
   type Dependency,
-  type DependencyTypeShort,
 } from '../../src/dependencies.js'
 import { Graph } from '../../src/graph.js'
 import { addNodes } from '../../src/ideal/add-nodes.js'
