+++ src/types/src/index.ts
@@ -368,3 +368,58 @@ export const assertPackument: (
 ) => asserts m is Packument = m => {
   asPackument(m)
 }
+
+/**
+ * Name of the package.json keys used to define different types of dependencies.
+ */
+export type DependencyTypeLong =
+  | 'dependencies'
+  | 'devDependencies'
+  | 'optionalDependencies'
+  | 'peerDependencies'
+
+/**
+ * Unique keys that define different types of dependencies relationship.
+ */
+export type DependencyTypeShort =
+  | 'dev'
+  | 'optional'
+  | 'peer'
+  | 'peerOptional'
+  | 'prod'
+
+/**
+ * A set of the possible long dependency type names,
+ * as used in `package.json` files.
+ */
+export const longDependencyTypes = new Set<DependencyTypeLong>([
+  'dependencies',
+  'devDependencies',
+  'peerDependencies',
+  'optionalDependencies',
+])
+
+/**
+ * A set of the short type keys used to represent dependency relationships.
+ */
+export const shortDependencyTypes = new Set<DependencyTypeShort>([
+  'prod',
+  'dev',
+  'peer',
+  'optional',
+  'peerOptional',
+])
+
+/**
+ * Maps between long form names usually used in `package.json` files
+ * to a corresponding short form name, used in lockfiles.
+ */
+export const dependencyTypes = new Map<
+  DependencyTypeLong,
+  DependencyTypeShort
+>([
+  ['dependencies', 'prod'],
+  ['devDependencies', 'dev'],
+  ['peerDependencies', 'peer'],
+  ['optionalDependencies', 'optional'],
+])
